Remove NODE_PATH hack for modules/ in favor of a custom module resolver
Summary:
Now that we've pulled in `babel-plugin-module-resolver`, this adds some code to read the `workspaces` field of the root `package.json` and adds aliases for each Yarn workspace. This provides a much cleaner workaround to the Yarn workspaces compatibility story than the existing `NODE_PATH` hack, which causes occasional issues when people accidentally have `nuclide-commons` installed in other places.

We have to be careful not to enable this when publishing modules, though, as those will be installed normally. Thus modules will be transpiled with `NUCLIDE_TRANSPILE_ENV` set to `production-modules`.

Reviewed By: matthewwithanm

Differential Revision: D7907223

fbshipit-source-id: a1a1332a3ad08128ed2e03211271fdf220711e8b