Nuclide HHVM Debugger: Work around segfault in HHVM on evaluate
Summary:
This diff adds a work around for a bug in the HHVM side of the debugger that causes a segfault on certain types of evaluations. The underlying bug is fixed in HHVM but this adds a mitigation until the next release is out.

Removing the context 'watch' or 'hover' from the evaluate commmand as a side effect that any errors generated by the watch or hover will be printed to the Nuclide console, whereas normally they are hidden from the user. This is preferable to crashing the webserver though.

Reviewed By: velocityboy

Differential Revision: D7974293

fbshipit-source-id: 7a6c1b6fc7b8c15ec03102ecf205c16311280461