Upgrade to babylon 7 (beta)
Summary:
`babylon@7` is technically still a beta pre-release but most of the tooling at Facebook has already started to adopt it.

This didn't require many changes - just a couple of AST node renames.

It looks like using "*" in plugins is no longer supported, so I enumerated the ones necessary in nuclide-rpc.

Reviewed By: wanderley

Differential Revision: D7976270

fbshipit-source-id: 90a30f70dbecf22b52b884b24814598784d96941