Use `hg root` for all CWD in test environment
Summary:
This is my attempt to unify all scripts to use the same `cwd` or `rootDir` in all files.
Right now it's completely random:
1. our tests path are relative to `fbsource`
2. we run most scripts from `fbsource/xplat`
3. testing cwd are `fbsource/xplat/js`
4. sandcastle job runs from `fbsource/fbandroid`
5. macOS based jobs run from `fbsource/objc`

In this diff i tried to use `fbsource/` in all paths that we interact with.

(also i changed `sandcastle.sh` accept optional config so it's not hardcoded in the file)

Reviewed By: mjesun

Differential Revision: D7852925

fbshipit-source-id: 4baf8cddc43d388269577a337cfe5f329c86d0f6