Extract all stateful HgService code into an HgRepositorySubscriptions class [1/n]
Summary:
This is the first step on a journey to make `HgService` a bag of functions rather than a remote object. Doing so will remove one of the last uses of remote constructors (cc semmypurewal).

Looking at the API of `HgService`, it seems that it can be cleanly separated into two parts: one that is fairly stateless (needing only the `workingDirectory` path) and another that wraps a set of Watchman subscriptions in order to observe repository state.

This diff extracts all the stateful subscription wrapping into a `HgRepositorySubscriptions` object, which can be separated from the other parts of `HgService`.

Reviewed By: tjfryan

Differential Revision: D7921682

fbshipit-source-id: 4818f75c7fb081a0efd5503e48cef0d7b9721820