Add HgService.createRepositorySubscriptions [2/n]
Summary:
The previous diff extracted all subscription-related code into `HgRepositorySubscriptions`. Now, this goes the extra step to decouple the subscriptions from `HgService`.

`HgRepositoryClient` now fetches the repository subscriptions class separately (via a method `createRepositorySubscriptions`).

This cleans up the need for `waitForWatchmanSubscriptions`, which is an unusual pattern. Instead, `createRepositorySubscriptions` only resolves after the Watchman subscription is complete (and errors if it fails).

`HgRepositoryClient` needed some slightly modification to handle the async-ness of `createRepositorySubscriptions`. I wrapped most of the previous observe calls in a `_tryObserve` wrapper that connects the subscription once it completes.

Reviewed By: tjfryan

Differential Revision: D7924111

fbshipit-source-id: 0da048590cfdc76433dd3d3e3616f3004f6adea9