Fix await vs .then inconsistencies in nuclide-open-files (and other cleanups)
Summary:
It looks like `await` and `.then` have different semantics in newer V8 versions: https://stackoverflow.com/questions/46254408/promise-resolution-order-and-await

`nuclide-open-files` actually cares about the ordering so we need to make sure we're consistent. It might be worth rewriting the code here to avoid the awaits everywhere (even with this change, we have to be careful to ensure that every file event goes through exactly one `await`).

Reviewed By: matthewwithanm

Differential Revision: D7872893

fbshipit-source-id: e82b3913b2ddd0f43fb5d7651873eac39fdf5339