Rename sample-nuclide-contributors / sample-ui-playground to dev-*
Summary:
This creates a new "dev" prefix that is also stripped in prod (via the `.npmignore`) but isn't disabled by default. So Nuclide contributors will get these in dev mode without having to dig through pages of settings!

I had to tell 2 people about this just today, which pushed me over the edge to do this.

I could also make these `sample-dev-*` and special case that in the `FeatureLoader` if that's preferable.

Reviewed By: matthewwithanm

Differential Revision: D7999653

fbshipit-source-id: ac63c5d709c460345f68643b0b79465337a4f600