Fix issue with early cancel of task runner registration
Summary:
The Buck integration test started failing after D7824569.

I managed to trace it back to this: fetching of task runner state is canceled if another task runner is registered while it's happening. This change allows them to run concurrently.

This is also the first recorded example in history of `switchMap()` *not* being the right operator :P

Reviewed By: stepanhruda

Differential Revision: D7999587

fbshipit-source-id: 5e60c33e1aaaa9975dcdeb7395ea29903085ecdf