prototype tunnel implementation in Big Dig
Summary:
Big Dig Tunnel Prototype

This is a work-in-progress with a couple goals:

1. Improve tunneling performance when compared to the Nuclide-RPC Tunnel
2. Better understand what it means for Big Dig to expose a tunneling API

Right now, the implementation supports bi-directional TCP tunnels from the big dig client to the big dig server, but not the other way around. I (think) I've modularized the code well enough to easily support tunneling in both directions as we move forward. I'll just need to add a couple RPC messages and tunnel ids (which could probably be constructed from the local/remote ports and the direction, or we could just use a UUID).

I currently implemented this as a Big Dig service, but I'm not sure that makes sense. If we move in this direction, it might be better to include this logic directly in the Client and the Server? My original implementation did this, but I decided I wanted to separate as much as I could from the Client and Server implementations.

The primary modules here are `Proxy.js` and `ConnectionManager.js`. The `Proxy` is the server that listens on the client for connections, and the `ConnectionManager` is the component that creates connections to the server running on the remote host. The communication between the two uses a poorly-thought-out-and-hacked-together RPC that uses JSON as the format. We can (and probably should) change this to avoid parsing and serializing JSON. At minimum, I probably should separate out the creation of the RPC messages as functions so that they can be changed in one place.

Last, but not least, the current implementation doesn't include the security mechanism we implemented for the Nuclide-RPC tunnel (we used a gk to restrict tunnels to ports that we deemed safe). There's a lot more work to be done in terms of security, but I'll defer that for now.

I'm mainly looking for general feedback relating to the components (e.g. the `Proxy` and the `ConnectionManager`). Are there components you think I'm missing? Does the code look reasonable (given that it's a rough draft, of course)? Any other thoughts before I move forward?

The next steps for me are to clean up the code a bit, and then compare the throughput with our Nuclide-RPC tunnel implementation to make sure we're heading in the right direction. Once I'm satisfied it's improving things, I'll implement the reverse tunnel API and start looking at how to hook this up in place of the Nuclide-RPC tunnel.

Reviewed By: hansonw

Differential Revision: D7602828

fbshipit-source-id: 527def467d834ecccac998b598cb18b690b0909b