move SocketManager handling into TunnelManager and create a TunnelManager on both ends
Summary: The idea here is to have a TunnelManager on both sides of the connection to support reverse tunneling. In this case, we've moved the SocketManager handling into TunnelManager, minimizing the code in launcher.js. But now we can create a Proxy and/or a SocketManager on either side, getting us closer to reverse tunnels.

Reviewed By: wanderley

Differential Revision: D7859923

fbshipit-source-id: 4989462be00cef41a82a05131d6060a77056468e