Add a "fork" option to createMultiLspLanguageService
Summary:
JavaScript-based LSP servers that are bundled with Nuclide only need an RPC component to resolve the JS path. Now that we're doing this in a bunch of places, it's worth adding a 'fork' flag to do this automatically. This makes it much easier to create these kinds of LSP servers.

To make this even easier, the next step would be to create a wrapper that takes care of the boilerplate parameters (like the file notifier and host services.)

Reviewed By: kcaze

Differential Revision: D8085166

fbshipit-source-id: 0e55732ade6fb294f348845cd82a7e653c66fba1