catch errors caused by trying to ping on an unopened socket
Summary:
This should catch a previously uncaught error that a few people have seen T29178119. It still logs the error.

Another way to fix this would be to check the readyState on the socket. I initially tried that, but flow couldn't find the readyState property. Is there somewhere that we spec out flow types for npm modules? I couldn't find an internal spec for ws$WebSocket. Maybe it's built in to flow?

Reviewed By: hansonw

Differential Revision: D8129485

fbshipit-source-id: 11d7bcb02e4829b45e35a0991eddb1cb7ce7101b