Defend against missing PID in process table
Summary:
rightparen added this a while ago but it broke health tracking on Windows, because we never added a "ps" equivalent for Windows - so all the process tables are empty in `getChildProcesses`. This breaks some `nullthrows` assumptions, which throws some uncaught exceptions.

It seems like a good idea to be generally defensive against the case where `process.pid` is missing from the process table, just in case?

Reviewed By: ebluestein

Differential Revision: D8132069

fbshipit-source-id: 87be3c4cb02e475cf321dee5e1b606a220b39a48