Support observeStatus in MultiProjectLanguageService and from LSP provider
Summary:
(1) This diff tells LSP servers that Nuclide supports the 'status' capability. If the LSP server sends an appropriate status message, then LspLanguageService will handle it correctly via the LanguageService 'observeStatus/clickStatus' APIs.

(2) This diff implements 'observeStatus/clickStatus' APIs correctly in MultiProjectLanguageService

On its own this diff doesn't do anything. But if an LSP server decides to use the capability that's offered, and if provider is written to surface observeStatus/clickStatus in the UI, then the user will see the status messages that the LSP server sends!

Reviewed By: kcaze

Differential Revision: D8124696

fbshipit-source-id: 18d6d4751c6d917e823a54972bb42ad5e030f19e