Nuclide debugger: try re-launching under sudo if attach fails
Summary:
When attempting to attach the debugger to a process that is running as root or another user, attach will fail with an error indicating the process could not be found. This diff adds one more attempt to launch the debugger under sudo before bailing with an error.

Note that this will only be successful in cases where the user is able to sudo, and sudo doesn't require the user to re-enter a password. (In those cases, we'll timeout and then fail to launch, as we did before this diff).

Reviewed By: velocityboy

Differential Revision: D8104935

fbshipit-source-id: e67277803dfb33e732266e749183cfbe0a6033fd