expose big dig client via getBigDigClientByNuclideUri
Summary:
expose big dig client from nuclide-remote-connection

This seems problematic to me. I feel like we need an async method/function to get the big dig client, since I could potentially call getOrCreate (which is async) to create the connection then subsequently call getBigDigClientByNuclideUri before it's actually created and added to the map. Unfortunately, since the server connection object doesn't get added to the map before initialization, it's currently impossible to know that there's a pending server connection for that host.

Interestingly, this seems like a problem for getOrCreate as well. If I call getOrCreate() on a host, and then immediately call getOrCreate on the same host before the first one is finished initializing, we'll have two server connection objects, one of which is inaccessible and floating around.

Am I wrong about this? I wonder if this approach is enough to get us up and running, or if we should seriously try to solve the problems I mentioned above.

Reviewed By: hansonw

Differential Revision: D8164873

fbshipit-source-id: 8a9c0f8cfbc63c4c0635b73e6bedc309a27fba38