Improve RPC call timeout message
Summary: Explicitly tell the user that this means their Nuclide server cannot be reached reliably. Usually this error means the network connection is down or spotty, this happens especially on poor WiFi or VPN. The "Failed to blah: Timeout after 60000..." message doesn't tell the user what's going on, and looks like an ordinary programming error.

Reviewed By: hansonw

Differential Revision: D8167786

fbshipit-source-id: c108a660fc514fef6e1ed8754f7a7a90d4b645ea