remove all event listeners from the socket once it's closed
Summary: This should remove some of the log spam, and it should resolve a memory leak (I think we were probably keeping the socket around despite removing the reference since we still had listeners attached). If this doesn't cause any problems, then we can clean up this code a lot (remove all the null checks on the socket in the event listeners).

Reviewed By: hansonw

Differential Revision: D8207289

fbshipit-source-id: 03cbfa45d73a8f72dadb23ed23f2361e5fe07364