Add FileSystemService.readdirSorted
Summary:
This is kind of an edge case, but for large directories the `sort()` call actually becomes a big bottleneck (10K files => 600ms to sort).

The solution is to offload the cost to the server in the form of a sorted `readdirSorted`. See the attached task for details.

Note: https://github.com/nodejs/node/issues/3232 indicates that on Linux/Mac `readdir` is actually sorted (but in a case-sensitive way). So we're almost there? If we're willing to relax the case restriction we could remove the sort on Linux/Mac.

Reviewed By: semmypurewal

Differential Revision: D8213348

fbshipit-source-id: 6798370565fb099944adbed550ca7dd230d8a6e2