Break long filenames at '/'
Summary: Previously, filenames would either be not wrapped at all or broken anywhere, depending on who was using the component. This change inserts zero with spaces at every '/' so that chrome can break the word at those points. The `(Added - 10 lines)` info after the filename is also now `inline-block` so it will not be wrapped halfway.

Reviewed By: tjfryan

Differential Revision: D8209455

fbshipit-source-id: 875dbca6ba258496c866ad75aafd38a94bf28e42