Scopes pane doesn't show any variables if its opened after the debugger is already paused
Summary:
The scopes pane can be opened after the debugger is paused if the user does debugger->show (any pane), or reset layout. Since it registers for notifications when the current call frame changes, if we're already paused, it doesn't get that notification and doesn't update scopes.

It should do an initial fetch as soon as its mounted.

Reviewed By: Goom11

Differential Revision: D8238336

fbshipit-source-id: aded6d379923da2150dad204ebdd0c0e6a0a1459