Use window/showStatus for status
Summary:
How should LSP server communicate status to the client?

In my first attempt I had it re-use the window/showMessageRequest method, via an optional field.

That was an ugly idea!

I think it's more in keeping with the spirit of LSP to use a method specifically for this: window/showStatus.

(also, it will make it easier for me to write integration tests for my LSP server if status messages have their own distinct method name).

Reviewed By: kcaze

Differential Revision: D8215884

fbshipit-source-id: 331ad68e6f52be996d77688699ac7f0d1bdc3a69