Debugger: Fix multiple breakpoints in the UI for the same breakpoint
Summary:
If a breakpoint is calibrated such that it ends up with a different endline than the line the user clicked in the gutter to set the breakpoint, our de-duping logic is failing to match the breakpoint if the same source line is clicked in the gutter again.

This results in multiple entries in the breakpoint list for the same break location, and all but the first can't actually be removed. (Clicking them in the UI does nothing).

Reviewed By: mostafaeweda

Differential Revision: D8241369

fbshipit-source-id: 02467f57fdb3583cf8bc55a742d066e3f5284547