Move diff gutter git-diff off process
Summary:
This works, but I'm a little unsure of how the performance with this will compare to the previous impl.  The `spawn` is now offprocess, but there's a lot more sending of large strings (for some definition of 'large') across IPC which seems like it could result in performance problems.

I don't notice any performance changes when testing the feature (either for better or for worse)

Reviewed By: hansonw

Differential Revision: D8242780

fbshipit-source-id: 3995a1834941ec9f37e53021a508f8d249b6a4fe