abstract encoding/decoding out of inline code
Summary:
I wanted to play around with binary formats (specifically msgpack and cbor) in the tunnel this week. They are basically drop-in replacements for JSON.stringify and JSON.parse, but with the advantage that they don't turn Buffers into larger array objects the way JSON.stringify and JSON.parse do. In trying this, I abstracted out the encoding/decoding of objects (including the base64 encoding) into its own submodule so I could just replace the encoder.

My experiments with msgpack were pretty disappointing, but I think it's to be expected. Unfortunately, since our lower level infra (QAT and WST) all expect strings, we end up having to base64 encode the binary data anyway. This, along with doing the encoding in JS land, kills performance.

Overall, we lose about 40% throughput by moving to msgpack. But maybe if we change the stack all the way down to support binary, we might see some wins? Not sure, but it's probably worth trying. We should at least try this in Big Dig land to see if we can make it work. Unfortunately we can't use native encoders if we're potentially targeting browsers.

Anyhow, that's a really long way of saying that I felt it was worthwhile to keep the encoding abstraction around. So here it is.

Reviewed By: hansonw

Differential Revision: D8239425

fbshipit-source-id: df02080dfa3a7e32bcd198aaf18e94d3b9b48a37