Nuclide task runner: Move buttons to the left
Summary:
Per XFN user research feedback from @[100002436853051:yuanzi], the action buttons (build/run/test/debug) should be on the right on the task runner toolbar.  This is because currently the user needs to move back and forth both visually and with the mouse to operate the toolbar.  For example, select the toolbar (HHVM/buck/relay/whatever) on the left. Move about 3/4 of the way to the right to enter a target in the target box. Click the settings gear to the right of the target box to enter some settings. Move to the right again to select a device. Then move the mouse all the way to the left to find the "debug" button.

This diff moves things around so the user's flow should be strictly left to right when operating the toolbar.

Reviewed By: stepanhruda

Differential Revision: D8243231

fbshipit-source-id: 1def5773d76cca1d24a8e8d87c6bf6f49800aaad