Raise Atom warning if there's a buck error fetching targets
Summary:
Listing aliases on my devserver often fails because some permissions are messed up on my buckd files. But there are other reasons (our fault and not) this operation can fail. Currently the error is displayed in the combobox dropdown but it's basically impossible to read, and therefore impossible to troubleshoot.

This diff raises an atom error notification so I can see what went wrong, and hopefully unblock myself.

Reviewed By: stepanhruda

Differential Revision: D8244144

fbshipit-source-id: 2e63a2e8af4ade711fcceb4454005827040ae275