Nuclide HHVM Debugger: Include server log directly in bugs
Summary: When the bug nub is clicked, we were trying to create a paste from the server log instead of including it directly as a perf optimization (to avoid streaming the log to the Nuclide client over Nuclide RPC). But this isn't working in some cases - especially On Demand, the current user might not have a Phabricator token, and we're getting empty logs with error messages. Not very useful.

Reviewed By: Goom11

Differential Revision: D8265410

fbshipit-source-id: f3eea9d6baddec0c1e0fb9f250c0407f2d6d9d94