Ensure that definition previews only fetch definitions from DefinitionCache
Summary:
ljw1004 reported a bug where hyperclick is occasionally flaky with higher-latency hyperclick providers. I tracked this down to the cancellation behavior of `LspLanguageService`, which always cancels previous definition requests when a newer request comes in.

The problem is that we have two things asking for the definition - the actual hyperclick request and the hyperclick preview request. Unfortunately since the hyperclick preview request is debounced it always cancels the hyperclick request. We normally don't notice this because of the cache, but sometimes the definition preview request doesn't match up perfectly with the definition request.

The real fix is to use cancellation tokens throughout all the language services, but in the meantime a simple workaround is to ensure that the definition preview request only ever uses the cache. This should be safe because the definition preview request should always come after the original definition request, and if it doesn't match it should be safe to ignore the request.

This diff adds:

1) `DefinitionCache.getCached` which only uses the cached value, and uses it for `getPreview`
2) fixes a quirk in `DefinitionCache` where earlier requests may clear the cached results of later requests, which helps ensure that definition previews are actually successful.

I noticed another bug where the datatip request may actually have the wrong position due to the fact that Chrome delays mousemove events until the next animation frame, but that's kind of a separate fix.

Reviewed By: ljw1004

Differential Revision: D8268052

fbshipit-source-id: 4e0c726938ae181621ea0dedc7efdd0e856f27d4