Better error reporting when the remote server command is invalid
Summary:
I noticed that if the remote server command is invalid we currently just display that the server "failed to start", which isn't very helpful when it comes to debugging (ideally we want the stdout of the command or something like that.)

I found out that `SshHandshake` isn't reporting the right error back to the `SshConnectionDelegate` - it unwraps the `innerError` but returns the wrapped one. I suspect this bug was due to the weird naming of the variables so I fixed that.

Also, in Nuclide's notification window, I made it so that 'failed to start' errors also print the original error message in the redbox.

Reviewed By: semmypurewal

Differential Revision: D8270488

fbshipit-source-id: 9e846ea108bd4a5e504214de28ce5084d3b30031