BreakpointListComponent Nit: SetState called in constructor
Summary:
The constructor creates an observable that calls set state. Moved this to componentDidMount so that setState doesn't get called until the component has mounted.

Note: kevinatorchen, if you run atom from the command line with the dev flag, `atom --dev`, then these React warnings show up in the console. It's just a warning, it shouldn't affect the behavior of your code.

Reviewed By: ebluestein

Differential Revision: D8268064

fbshipit-source-id: 0d3835a70b4237ac19d6c032076a32f953299ba1