Nuclide IDE returns position after the cursor according to LSP
Summary:
Nuclide currently returns the position of the character last typed, but it should return the position of the cursor (the position following the last character typed).

The dependent diff modifies the hack server to account for this change. To avoid breaking the IDE, it first checks if the current position points to a trigger character and if not, returns the previous position.

Reviewed By: hansonw

Differential Revision: D8227027

fbshipit-source-id: 81ac012f137329b66f047784781de8ba4c647b89