introduce a boolean switch for new and old tunnel implementations
Summary:
This introduces a boolean variable that controls whether tunnel creation is happening via Nuclide-RPC or BigDig. Right now it defaults to false.

There are a few differences between the new implementation and the old one that may affect users. Right now, we're not plumbing through the 'connection' events, so we never change the state from 'ready' to 'active'. This isn't a ton of work, but it's not a priority right now.

We're also never throwing the error event from the Tunnel (although we are handling it once it gets thrown). That's something that we should fix before we start rolling this out.

Based on the way we wrote the new tunnel implementation, adding reference counting and sharing tunnels is trivial, so we'll probably add that functionality pretty soon since we'll need it for other big dig clients outside of Nuclide (VSCode, specifically).

Next steps are to put the flag behind a gk, make sure we're sending errors correctly, roll it out, monitor it, then code clean-up.

Reviewed By: stepanhruda

Differential Revision: D8224000

fbshipit-source-id: 249d0863c278437f475370eedf98d388b8fe3018