Base outline on range-inclusion not containerName
Summary:
How should we construct an outline view out of the flat list textDocument/documentSymbols?

When I first wrote LspLanguageService I believed the "range" in documentSymbols referred to the range of the identifier rather than the range of the entire construct (e.g. in `class C {...}` I believed it referred to the range of the identifier `C` rather than the range of the whole thing). Therefore we couldn't reconstruct hierarchy based on range inclusion. I used containerName instead.

But it appears that standard industry practice is that "range" is indeed the range of the whole thing. Therefore we can use range inclusion to reconstruct the outline.

This makes Nuclide happier! Two features become lit up if we provide the range of the entire construct -- (1) breadcrumb bar, (2) highlight where you are inside the outline view.

NOTE: I can't land this until Hack has been changed to emit ranges correctly.

Reviewed By: kcaze

Differential Revision: D8178812

fbshipit-source-id: 38d1f41cc55de0b41a478b5365451242094631f1