Add nuclide-onboarding-core-tasks package and create recent IDE experience task
Summary: Created the nuclide-onboarding-core-tasks package, which provides an OnboardingFragment for the Editor/IDE Experience task. Modified the Onboarding main.js to use a Model with state state keeps track of the current activeTask and all of the provided tasks. Bind the Model.toObservable as props to the OnboardingPaneItem so that the OnboardingPaneItem re-renders on Model changes.

Reviewed By: matthewwithanm

Differential Revision: D8128378

fbshipit-source-id: f52363ffc7f66d076c622b3a9e57c52aa3b41126