Debugger autogen: Support a Path type that expands ~/ to home directory
Summary:
Some inputs in the debugger configuration dialog are paths, which users expect to be able to type ~/ in to expand to the home directory. This needs to be resolved on the machine that targetUri refers to though, not on the client.

This diff adds a path configuration type. Any path has ~/ expanded by the RPC framework (if available) before passing it on to the debugger arguments.

Reviewed By: Goom11

Differential Revision: D8269225

fbshipit-source-id: 0bf0de83b4386c9c27f418c31cc2bf67720c0bfb