Track timeouts for e2e-autocomplete
Summary:
This was a mistake - the data needs to include timeouts or else it doesn't provide an accurate picture of the user experience.
It's possible that multiple providers timeout so it's hard to attribute these to a single "slowestProvider" - but this logs the number of pendingProviders when the request timed out.

Reviewed By: wanderley

Differential Revision: D8291880

fbshipit-source-id: e82cf5e42d9f3a40fbbc10d8052548cd738cba6b