Record latency of successful go-to-definition clicks
Summary:
Similar to how we were able to determine that 500ms was a reasonable cutoff for autocomplete requests, we should do the same kind of analysis for go-to-definition events. By logging how long a user waited before clicking, we should be able to determine a good target for go-to-definition latency.

We don't really need to enforce a cutoff in this case, but it'll be really good to understand how long users are willing to wait before they can click on a result.

Reviewed By: wanderley

Differential Revision: D8292092

fbshipit-source-id: 11e58acefe9e09d463c8d73d35b734938b79bcb0