Mount repo root when loading remote projects
Summary:
Instead of mounting each of the directories in the project file's `paths`, we'll mount the repo root. This means other files in the repo will still be exposed via search, etc. Ideally, we would mount the `paths` and it would be up to the other code to determine if it wanted to use the mounted directories or repo roots, however, for now at least, this is an easier change to make.

Note that this functionality is extracted into `ProjectManager.open()` in D8329525.

Reviewed By: hansonw

Differential Revision: D8172839

fbshipit-source-id: d2c706187fe6b90e9e4ed11773b7963a303a717b