Attempt plug diff gutters memory leak
Summary:
I initially didn't want to `takeUntil` the fetchFileContentsForHead because it buffers and I was afraid that closing the wrong editor at the wrong time would prevent all the files whose contents were buffered with it from being fetched, and I didn't think of putting the takeUntil anywhere earlier than the end of an Observable operation chain.

Where I think this runs into a memory leak is if a text editor never becomes visible before it's closed, causing the observable that starts with `observePaneItemVisibility` to never terminate (despite having a `.first()` in it).

Assuming I understand `takeUntil` correctly, this should clean up the reference to the textEditor when it is closed but prevent it from cancelling an in-flight `hg cat` operation.

Reviewed By: hansonw

Differential Revision: D8370547

fbshipit-source-id: b16396118d242ee9e5102ea1ed4f110547f78691